feat: migrate project from JavaScript to TypeScript
