+++ services/email-service/package.json
@@ -3,19 +3,26 @@
   "version": "1.0.0",
   "description": "",
   "main": "index.js",
+  "type": "commonjs",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "dev": "nodemon index.js"
+    "build": "tsc",
+    "start": "ts-node index.ts",
+    "dev": "nodemon --ext ts --exec ts-node index.ts"
   },
   "keywords": [],
   "author": "",
   "license": "ISC",
   "dependencies": {
+    "@types/node": "^22.13.0",
+    "@types/nodemailer": "^6.4.17",
     "dotenv": "^16.4.7",
     "moleculer": "^0.14.35",
     "moleculer-web": "^0.10.7",
     "nats": "^2.29.1",
     "nodemailer": "^6.10.0",
-    "nodemon": "^3.1.9"
+    "nodemon": "^3.1.9",
+    "ts-node": "^10.9.2",
+    "typescript": "^5.7.3"
   }
 }
