--- services/email-service/src/events/handleUserCreated.js
@@ -1,11 +1,0 @@
-const { errorHandler } = require("../helpers/error-handler");
-
-module.exports = {
-    'user.created'(ctx){
-        try{
-            return ctx.call('email.sendGreetingEmail',ctx.params);
-        }catch(err){
-            return errorHandler(err);
-        }
-    }
-}
