+++ services/email-service/src/events/handleUserCreated.ts
@@ -1,0 +1,12 @@
+import errorHandler from '../helpers/error-handler';
+import { Context } from 'moleculer';
+import { UserData } from '../types/user.types';
+
+export default async function handler(ctx: Context<UserData>){
+    try{
+        const result = await ctx.call('email.sendGreetingEmail', ctx.params);
+        return result;
+    }catch(err){
+        return errorHandler(err);
+    }
+}
