--- services/email-service/src/helpers/email-helper.js
@@ -1,9 +1,0 @@
-module.exports = {
-    getGreetingEmailSubject(){
-        return 'Welcome to Our Platform';
-    },
-    
-    getGreetingEmailBody(input){
-        return `Hello ${input.firstName},\n\nWelcome to our platform!\n\nBest Regards,\nTeam`;
-    }
-}
