+++ services/email-service/src/helpers/email-helper.ts
@@ -1,0 +1,9 @@
+import { UserData } from "../types/user.types";
+
+export function getGreetingEmailSubject(): string{
+    return 'Welcome to Our Platform';
+};
+    
+export function getGreetingEmailBody(input: UserData): string{
+    return `Hello ${input.firstName},\n\nWelcome to our platform!\n\nBest Regards,\nTeam`;
+};
