+++ services/email-service/src/types/error.handler.types.ts
@@ -1,0 +1,9 @@
+export interface ErrorResponse {
+    success: boolean;
+    error: {
+        message: string;
+        code: number;
+        type: string;
+    },
+    statusCode: string;
+}
