+++ services/email-service/src/types/nodemailer.types.ts
@@ -1,0 +1,6 @@
+export interface MailOptions {
+    from: string|undefined;
+    to: string;
+    subject: string;
+    text: string;
+}
