+++ services/email-service/src/types/user.types.ts
@@ -1,0 +1,5 @@
+export interface UserData {
+    firstName: string;
+    lastName: string;
+    email: string;
+}
