--- services/user-service/src/events/userCreated.js
@@ -1,5 +1,0 @@
-module.exports = {
-    'user.created'(payload){
-        console.log(`Event triggered: User ${payload.firstName} created`);
-    }
-}
