+++ services/user-service/src/events/userCreated.ts
@@ -1,0 +1,5 @@
+import { UserInput } from "../types/user.types";
+
+export default function handler(payload: UserInput): void{
+    console.log(`Event triggered: User ${payload.firstName} created`);
+}
