+++ services/user-service/src/types/user.types.ts
@@ -1,0 +1,5 @@
+export interface UserInput {
+    firstName: string;
+    lastName: string;
+    email: string;
+}
